better organize parsers and related custom exceptions
