+++ src/main/java/net/querz/mcaselector/ui/dialog/FilterChunksDialog.java
@@ -23,8 +23,8 @@ import net.querz.mcaselector.filter.GroupFilter;
 import net.querz.mcaselector.filter.Operator;
 import net.querz.mcaselector.filter.XPosFilter;
 import net.querz.mcaselector.filter.ZPosFilter;
-import net.querz.mcaselector.headless.FilterParser;
-import net.querz.mcaselector.headless.ParseException;
+import net.querz.mcaselector.filter.FilterParser;
+import net.querz.mcaselector.exception.ParseException;
 import net.querz.mcaselector.text.Translation;
 import net.querz.mcaselector.ui.GroupFilterBox;
 import net.querz.mcaselector.ui.UIFactory;
